@@ -18,7 +18,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ^3.0.0
 
 dependencies:
   flutter:
@@ -26,45 +26,50 @@ dependencies:
 
   # UI
   cupertino_icons: ^1.0.4
-  flutter_svg: ^1.0.3
+  flutter_svg: ^2.0.7
   nstack:
     git:
       url: https://github.com/nstack-io/flutter-sdk.git
       ref: v0.5.1
 
   # Net
-  dio: ^4.0.6
+  dio: ^5.1.2
   jaguar_jwt: ^3.0.0
 
   # Utility
-  device_info_plus: ^3.2.3
-  package_info_plus: ^1.4.2
+  device_info_plus: ^9.0.0
+  package_info_plus: ^4.0.0
 
   # Persistence
-  shared_preferences: ^2.0.13
   flutter_secure_storage: ^7.0.1
+  shared_preferences: ^2.1.1
 
   # Architecture
-  freezed_annotation: 2.0.1
-  json_annotation: ^4.4.0
-  injectable: 1.5.3
-  get_it: ^7.2.0
-  auto_route: ^3.2.4
-  flutter_bloc: ^8.0.1
-  rxdart: ^0.27.3
+  freezed_annotation: ^2.2.0
+  json_annotation: ^4.8.1
+  injectable: ^2.1.1
+  get_it: ^7.6.0
+  auto_route: ^7.1.0
+  flutter_bloc: ^8.1.2
+  rxdart: ^0.27.7
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   mocktail: ^0.3.0
-  build_runner: ^2.1.11
-  injectable_generator: 1.5.3
-  freezed: 2.0.2
-  json_serializable: ^6.1.5
-  auto_route_generator: ^3.2.3
-  flutter_lints: ^1.0.4
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
+  build_runner: ^2.4.6
+  injectable_generator: ^2.1.5
+  freezed: ^2.3.3
+  json_serializable: ^6.6.2
+  auto_route_generator: ^7.0.0
+  flutter_lints: ^2.0.1
+
+  # Simplifed work with assets
+  flutter_gen_runner: ^5.3.1
+
+flutter_gen:
+  integrations:
+    flutter_svg: true
 
 # The following section is specific to Flutter.
 flutter:
