@@ -1,6 +1,6 @@
 import 'package:flutter_template/injection/config_module.dart';
 import 'package:flutter_template/presentation/app_flavor.dart';
-import 'package:flutter_template/presentation/routes/router.gr.dart';
+import 'package:flutter_template/presentation/routes/router.dart';
 
 import 'injector.dart';
 
